faiss/requirements.txt
langchain
langchain-google-genai
langchain_community
python-dotenv
faiss-cpu
gradio